Backend(controllers): rename Apis to Controllers, remove unused controllers
