+++ packages/backend/src/middlewares/async.js
@@ -1,0 +1,3 @@
+export default (fn) => (req, res, next) => {
+  Promise.resolve(fn(req, res, next)).catch(next);
+};
