+++ packages/backend/src/routes.js
@@ -1,21 +1,14 @@
-import Index from "apis/IndexApi";
-import User from "apis/UserApi";
-import Application from "apis/ApplicationApi";
-import Network from "apis/NetworkApi";
+import Application from "controllers/ApplicationController";
+import Index from "controllers/DefaultController";
+import Network from "controllers/NetworkController";
+import User from "controllers/UserController";
 
-/**
- * @param {object} expressApp Express application object.
- */
 export function configureRoutes(expressApp) {
-  // Index API
   expressApp.use("/", Index);
 
-  // Users API
   expressApp.use("/api/users", User);
 
-  // Applications API
   expressApp.use("/api/applications", Application);
 
-  // Network API
   expressApp.use("/api/network", Network);
 }
